--- packages/server/modules/comments/services/retrieval/deprecated/index.ts
@@ -1,1 +1,0 @@
-export { getComments } from './getComments'
